Make GC storage agnostic
