+++ src/peergos/server/Main.java
@@ -465,7 +465,7 @@ public class Main {
                 boolean enableGC = a.getBoolean("enable-gc", false);
                 ContentAddressedStorage.HTTP ipfs = new ContentAddressedStorage.HTTP(ipfsApi, false);
                 if (enableGC) {
-                    GarbageCollector gced = new GarbageCollector(ipfs, a.getInt("gc.period.millis", 60 * 60 * 1000));
+                    IpfsGarbageCollector gced = new IpfsGarbageCollector(ipfs, a.getInt("gc.period.millis", 60 * 60 * 1000));
                     gced.start();
                     localDht = new CachingStorage(gced, dhtCacheEntries, maxValueSizeToCache);
                 } else
