+++ src/peergos/server/storage/DeletableContentAddressedStorage.java
@@ -1,0 +1,13 @@
+package peergos.server.storage;
+
+import peergos.shared.io.ipfs.multihash.*;
+import peergos.shared.storage.*;
+
+import java.util.stream.*;
+
+public interface DeletableContentAddressedStorage extends ContentAddressedStorage {
+
+    Stream<Multihash> getAllFiles();
+
+    void delete(Multihash hash);
+}
